requirements.txt
# requirements.txt

Flask
flask-cors
gunicorn
easyocr
opencv-python-headless
numpy
Pillow